requirements.txt
aiodataloader==0.2.1
aniso8601==7.0.0
asgiref==3.5.2
colorful-print==0.0.6
Django==3.1.5
django-filter==2.3.0
djangorestframework==3.13.1
Faker==14.1.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
python-dateutil==2.8.2
pytz==2022.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2022.1
